# microservices/legal-advogado/swagger.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [
    { "url": "http://localhost:8080", "description": "Generated server url" }
  ],
...
  "paths": {
...
    "/api/v1/agendamentos/avaliar": {
      "post": {
        "tags": ["Agendamentos"],
        "summary": "Avalia uma consulta",
        "description": "Permite a um cliente fornecer uma nota e um comentário para uma consulta já realizada.",
        "operationId": "avaliarConsulta",
        "parameters": [
          {
            "name": "agendamentoId",
            "in": "query",
            "description": "ID do agendamento a ser avaliado",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "clienteId",
            "in": "query",
            "description": "ID do cliente que está avaliando",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "rating",
            "in": "query",
            "description": "Nota da avaliação (1 a 5)",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          },
          {
            "name": "comentario",
            "in": "query",
            "description": "Comentário opcional sobre a avaliação",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Consulta avaliada com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseBoolean" }
              }
            }
          },
          "400": {
            "description": "Dados de requisição inválidos (ex: nota fora do intervalo)",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseBoolean" }
              }
            }
          },
          "404": {
            "description": "Agendamento não encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseBoolean" }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseBoolean" }
              }
            }
          }
        }
      }
    },
    "/api/v1/agendamentos/agendar": {
      "post": {
        "tags": ["Agendamentos"],
        "summary": "Agenda uma nova consulta",
        "description": "Permite a um cliente agendar uma consulta com um profissional, verificando disponibilidade e enviando notificações.",
        "operationId": "agendarConsulta",
        "parameters": [
          {
            "name": "profissionalId",
            "in": "query",
            "description": "ID do profissional para o agendamento",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "clienteId",
            "in": "query",
            "description": "ID do cliente que está agendando",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "horarioAgendamento",
            "in": "query",
            "description": "Horário desejado para a consulta (formato ISO 8601, ex: 2024-07-25T10:00:00)",
            "required": true,
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "tipoServico",
            "in": "query",
            "description": "Tipo de serviço a ser agendado (ex: 'Consulta Online', 'Reunião Presencial')",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "201": {
            "description": "Consulta agendada com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseBoolean" }
              }
            }
          },
          "400": {
            "description": "Dados de requisição inválidos (ex: horário indisponível)",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseBoolean" }
              }
            }
          },
          "404": {
            "description": "Profissional não encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseBoolean" }
              }
            }
          },
          "409": {
            "description": "Conflito de agendamento (horário já ocupado)",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseBoolean" }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseBoolean" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/areas-atuacao": {
      "get": {
        "tags": ["Áreas de Atuação"],
        "summary": "Lista todas as áreas de atuação",
        "description": "Retorna uma lista de todas as áreas de atuação cadastradas.",
        "operationId": "getAllAreasAtuacao",
        "responses": {
          "200": {
            "description": "Áreas de Atuação listadas com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListAreaAtuacaoResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListAreaAtuacaoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/areas-atuacao/{id}": {
      "get": {
        "tags": ["Áreas de Atuação"],
        "summary": "Busca uma área de atuação pelo ID",
        "description": "Retorna os detalhes de uma área de atuação específica.",
        "operationId": "getAreaAtuacaoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da Área de Atuação",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Área de Atuação encontrada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseAreaAtuacaoResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Área de Atuação não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseAreaAtuacaoResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseAreaAtuacaoResponseDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AreaAtuacaoResponseDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "descricao": { "type": "string" }
        }
      },
...
      "BaseResponseBoolean": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["SUCESSO", "ERRO", "AVISO"] },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": { "type": "boolean" },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "BaseResponseListAreaAtuacaoResponseDTO": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["SUCESSO", "ERRO", "AVISO"] },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AreaAtuacaoResponseDTO" }
          },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
      "BaseResponseAreaAtuacaoResponseDTO": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["SUCESSO", "ERRO", "AVISO"] },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": { "$ref": "#/components/schemas/AreaAtuacaoResponseDTO" },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
...
    }
  }
}